[Design 🎨] Navbar Menu hover 애니메이션 변경
