+++ src/components/Navbar/components/NavbarRightList/NavbarLoggedInMenu/NavbarLoggedInMenu.tsx
@@ -1,6 +1,6 @@
 import { NavbarMenuProps } from "@/components/Navbar/types"
 import { NavbarMenuList } from "./../NavbarRightList.Styles"
-import NavbarMenuItem from "../NavbarMenuItem"
+import NavbarMenuItem from "../NavbarMenuItem/NavbarMenuItem"
 
 const NavbarLoggedInMenu = ({
   handleMenuClick,
