+++ src/components/Navbar/components/NavbarRightList/NavbarMenuItem/NavbarMenuItem.tsx
@@ -1,0 +1,23 @@
+import Button from "@/components/Button/Button"
+import { theme } from "@/styles/theme"
+import { NavbarMenuProps } from "@/components/Navbar/types"
+import * as S from "./NavbarMenuItem.Styles"
+
+const NavbarMenuItem = ({ children, handleMenuClick }: NavbarMenuProps) => {
+  return (
+    <S.NavbarMenuContainer>
+      <Button
+        $fontColor={theme.colors.white}
+        onClick={() => {
+          // string | undefined 문제..
+          handleMenuClick(children as string)
+        }}
+      >
+        <p>{children}</p>
+      </Button>
+      <S.NavbarMenuDivider />
+    </S.NavbarMenuContainer>
+  )
+}
+
+export default NavbarMenuItem
